refactor(login): break down 133-line page into modular components (→32 lines)
- Extract useGithubLogin hook (35 lines): auth logic with error handling
- Extract LoginHeader component (27 lines): branding, badges, welcome text
- Extract ErrorAlert component (11 lines): error display with conditional rendering
- Extract LoginButtons component (50 lines): GitHub login and Learn More CTAs
- Extract TermsFooter component (24 lines): terms and privacy links
- Create components/index.ts for clean re-exports
- Main LoginPage reduced from 133→32 lines
- Improved testability and component reusability
- Follows established feature module pattern

Part of #203 (files over 200 lines reduction)
