+++ apps/registry/app/login/components/index.ts
@@ -1,0 +1,5 @@
+export { useGithubLogin } from './useGithubLogin';
+export { LoginHeader } from './LoginHeader';
+export { ErrorAlert } from './ErrorAlert';
+export { LoginButtons } from './LoginButtons';
+export { TermsFooter } from './TermsFooter';
