+++ apps/registry/app/login/page.tsx
@@ -1,40 +1,16 @@
 'use client';
 
-import { useState } from 'react';
-import { supabase } from '../lib/supabase';
-import { Badge, Button, Card, CardContent } from '@repo/ui';
-import { Github, FileJson, ArrowRight } from 'lucide-react';
-
-import Link from 'next/link';
-
-// Get the app URL from environment, fallback to window.location.origin
-const getAppUrl = () => {
-  if (typeof window === 'undefined') return '';
-  return process.env.NEXT_PUBLIC_APP_URL || window.location.origin;
-};
+import { Card, CardContent } from '@repo/ui';
+import {
+  useGithubLogin,
+  LoginHeader,
+  ErrorAlert,
+  LoginButtons,
+  TermsFooter,
+} from './components';
 
 export default function LoginPage() {
-  const [error, setError] = useState<string | null>(null);
-
-  const handleGithubLogin = async () => {
-    try {
-      const { error } = await supabase.auth.signInWithOAuth({
-        provider: 'github',
-        options: {
-          scopes: 'read:user gist',
-          queryParams: {
-            access_type: 'offline',
-            prompt: 'consent',
-          },
-          redirectTo: `${getAppUrl()}/editor`,
-        },
-      });
-
-      if (error) throw error;
-    } catch (error) {
-      setError(error instanceof Error ? error.message : 'An error occurred');
-    }
-  };
+  const { error, handleGithubLogin } = useGithubLogin();
 
   return (
     <div className="min-h-screen flex items-center justify-center bg-gray-50 py-12 px-4 sm:px-6 lg:px-8 relative overflow-hidden">
@@ -45,87 +21,10 @@ export default function LoginPage() {
 
       <Card className="max-w-md w-full relative backdrop-blur-xl bg-white/80 border-none shadow-xl">
         <CardContent className="p-8 space-y-6">
-          <div className="text-center space-y-2">
-            <div className="flex justify-center mb-6">
-              <Badge className="animate-pulse" variant="secondary">
-                <span className="w-2 h-2 rounded-full bg-green-400 animate-pulse mr-2" />
-                Secure GitHub Login
-              </Badge>
-            </div>
-            <div className="flex items-center justify-center gap-2 mb-2">
-              <FileJson className="w-8 h-8 text-primary" />
-              <h2 className="text-3xl font-bold tracking-tight text-gray-900">
-                JSON Resume
-              </h2>
-            </div>
-            <h3 className="text-xl text-gray-600">Welcome Back</h3>
-            <p className="text-gray-600 max-w-sm mx-auto">
-              Sign in to manage your resume, explore themes, and share your
-              professional profile.
-            </p>
-          </div>
-
-          {error && (
-            <div className="bg-red-50 p-4 rounded-md">
-              <p className="text-sm text-red-700">{error}</p>
-            </div>
-          )}
-
-          <div className="space-y-4">
-            <Button
-              onClick={handleGithubLogin}
-              size="lg"
-              className="w-full gap-2 text-lg h-14 hover:scale-105 transition-transform group"
-            >
-              <Github className="w-5 h-5" />
-              Continue with GitHub
-              <ArrowRight className="w-4 h-4 ml-2 group-hover:translate-x-1 transition-transform" />
-            </Button>
-
-            <div className="text-center space-y-4">
-              <div className="relative">
-                <div className="absolute inset-0 flex items-center">
-                  <div className="w-full border-t border-gray-200"></div>
-                </div>
-                <div className="relative flex justify-center text-sm">
-                  <span className="px-2 bg-white/80 text-gray-500">
-                    New to JSON Resume?
-                  </span>
-                </div>
-              </div>
-
-              <Button
-                variant="outline"
-                size="lg"
-                className="w-full gap-2 text-lg h-14 hover:scale-105 transition-transform group"
-                asChild
-              >
-                <Link href="/">
-                  Learn More
-                  <ArrowRight className="w-4 h-4 ml-2 group-hover:translate-x-1 transition-transform" />
-                </Link>
-              </Button>
-            </div>
-          </div>
-
-          <div className="text-center">
-            <p className="text-sm text-gray-600">
-              By signing in, you agree to our{' '}
-              <Link
-                href="/terms"
-                className="font-medium text-gray-900 hover:text-primary transition-colors"
-              >
-                Terms of Service
-              </Link>{' '}
-              and{' '}
-              <Link
-                href="/privacy"
-                className="font-medium text-gray-900 hover:text-primary transition-colors"
-              >
-                Privacy Policy
-              </Link>
-            </p>
-          </div>
+          <LoginHeader />
+          <ErrorAlert error={error} />
+          <LoginButtons onGithubLogin={handleGithubLogin} />
+          <TermsFooter />
         </CardContent>
       </Card>
     </div>
